Drop connection into once FIN packet received (and summary written)
Changed the script to reduce memory use so it can be run for longer
periods of time without bloating.  It deletes info about a connection
after the FIN packet is seen, since it doesn't need to know about it
after that.

The code in the original script that displays a summary of data
transfers by IP address was removed from this version as a result.